Combine components from developers into one scene
New scene will be primary scene for build.
Added functionality to menu so game starts from the menu, as well as the
option to pause and quit to menu.

BUGS:
need to remove the bullet shells on LeaveGame()
player needs to be disabled on pause
killing enemies causes navmesh errors